Move stake stats to redux; fix re-rerender on top menu of rose price
