+++ src/pages/App.tsx
@@ -20,6 +20,7 @@ import Swap from "./Swap"
 import Web3ReactManager from "../components/Web3ReactManager"
 import Withdraw from "./Withdraw"
 import fetchFarmStats from "../utils/fetchFarmStats"
+import fetchStakeStats from "../utils/fetchStakeStats"
 import fetchSwapStats from "../utils/getSwapStats"
 import fetchTokenPricesUSD from "../utils/updateTokenPrices"
 import { useActiveWeb3React } from "../hooks"
@@ -110,9 +111,13 @@ function GasAndTokenPrices({
   const fetchAndUpdateFarmStats = useCallback(() => {
     void fetchFarmStats(dispatch)
   }, [dispatch])
+  const fetchAndUpdateStakeStats = useCallback(() => {
+    void fetchStakeStats(dispatch)
+  }, [dispatch])
   // usePoller(fetchAndUpdateGasPrice, 5 * 1000)
   usePoller(fetchAndUpdateTokensPrice, BLOCK_TIME * 3)
   usePoller(fetchAndUpdateSwapStats, BLOCK_TIME * 280) // ~ 1hr
   usePoller(fetchAndUpdateFarmStats, BLOCK_TIME * 70) // ~ 15min
+  usePoller(fetchAndUpdateStakeStats, BLOCK_TIME * 70)
   return <>{children}</>
 }
